+++ crypto/src/sumcheck/mod.rs
@@ -1,1 +1,2 @@
-pub mod protocol;
+pub mod prover;
+pub mod verifier;
