+++ crypto/src/sumcheck/verifier.rs
@@ -1,56 +1,12 @@
-// sumcheck.rs (refactored version)
-
 use crate::fiat_shamir::default_transcript::DefaultTranscript;
 use crate::fiat_shamir::is_transcript::IsTranscript;
 use alloc::vec::Vec;
 use lambdaworks_math::field::element::FieldElement;
 use lambdaworks_math::field::traits::IsField;
 use lambdaworks_math::polynomial::{
-    dense_multilinear_poly::DenseMultilinearPolynomial,
-    Polynomial, // Represents univariate polynomials.
+    dense_multilinear_poly::DenseMultilinearPolynomial, Polynomial,
 };
 use lambdaworks_math::traits::ByteConversion;
-
-/// Prover for the Sum-Check protocol using DenseMultilinearPolynomial.
-pub struct Prover<F: IsField>
-where
-    <F as IsField>::BaseType: Send + Sync,
-{
-    pub poly: DenseMultilinearPolynomial<F>,
-    pub claimed_sum: FieldElement<F>,
-    pub current_round: usize,
-}
-
-impl<F: IsField> Prover<F>
-where
-    <F as IsField>::BaseType: Send + Sync,
-{
-    pub fn new(poly: DenseMultilinearPolynomial<F>) -> Self {
-        let evals = poly.to_evaluations();
-        let claimed_sum = evals.into_iter().sum();
-        Self {
-            poly,
-            claimed_sum,
-            current_round: 0,
-        }
-    }
-
-    pub fn c_1(&self) -> FieldElement<F> {
-        self.claimed_sum.clone()
-    }
-
-    /// Receives the challenge \( r_j \) from the verifier, fixes the last variable to that value,
-    /// and returns the univariate polynomial for the next variable.
-    /// (The univariate polynomial is represented as a Polynomial with two coefficients.)
-    pub fn round(&mut self, r_j: FieldElement<F>) -> Polynomial<FieldElement<F>> {
-        // Fix the last variable (using the integrated method in dense_multilinear_poly)
-        self.poly = self.poly.fix_last_variable(&r_j);
-        // Obtain the univariate polynomial: sum of evaluations with the last variable fixed to 0 and 1.
-        let univar = self.poly.to_univariate();
-        self.current_round += 1;
-        univar
-    }
-}
 
 /// Result of a verifier round.
 pub enum VerifierRoundResult<F: IsField>
@@ -165,12 +121,11 @@ where
 
 /// Channel trait for the transcript used in verification.
 pub trait Channel<F: IsField> {
-    /// Appends data (e.g., a field element) to the transcript.
     fn append_field_element(&mut self, element: &FieldElement<F>);
-    /// Draws a challenge from the field based on the current transcript state.
     fn draw_felt(&mut self) -> FieldElement<F>;
 }
 
+// Implementation of Channel for your DefaultTranscript
 impl<F> Channel<F> for DefaultTranscript<F>
 where
     F: IsField,
@@ -187,6 +142,8 @@ where
 
 #[cfg(test)]
 mod sumcheck_tests {
+    use crate::sumcheck::prover::Prover;
+
     use super::*;
     use lambdaworks_math::field::fields::u64_prime_field::U64PrimeField;
